backend/ml-models/requirements.txt
transformers>=4.55.4
torch>=2.6.0
Pillow>=10.2.0
numpy>=2.0.2
opencv-python>=4.9.0.80
scikit-learn>=1.4.0
